requirements.txt
Django==1.11
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.4
